fix(metrics): address review on the session-type breakdown
- session-changes: unsubscribing the last viewer cleared the coalescing
  timer that in-process watchers also wait on, so a pending
  session-directory report was dropped and the change waited for an
  unrelated notify. The timer is now cleared only when no watcher is left
  either; the on-demand hooks stay tied to the viewer count alone.
- session-directory: the stored row is parsed into domain values in the
  repository rather than the service, which is where translating a
  persistence representation belongs — and parsing straight into the
  contract's own mode schema drops both enum casts.
- features: the page claimed a flag never gates the underlying API, which
  this feature's server-side rollup gate contradicted. It now describes
  that third kind of gate and what still holds: the procedure stays open,
  the field comes back empty.
- runtime-delivery: the page described the channel as two routes with
  hello as the only agent-initiated call. The session-directory report is
  a third, so it is named there and in the outbound-channel invariant.

Left as-is: the empty-session-id rows the reviewer wanted filtered out of
the rollup. Excluding them would drop their cost from the breakdown while
the total above it still counts them; booking them as unattributed is
what keeps the two consistent.
